+++ src/import-organizer/import-grouping/ImportGroupKeyword.ts
@@ -1,12 +1,12 @@
 /**
  * Enum for the different special keywords of the KeywordImportGroup.
- * 
+ *
  * @export
  * @enum {number}
  */
 export enum ImportGroupKeyword {
-    Modules,
-    Plains,
-    Workspace,
-    Remaining,
+  Modules,
+  Plains,
+  Workspace,
+  Remaining,
 }
