+++ src/import-organizer/import-grouping/RemainImportGroup.ts
@@ -1,35 +1,35 @@
 import { Import, StringImport } from 'typescript-parser';
 
-import { importSort } from '../utilities/utilityFunctions';
+import { importSort } from '../../utilities/utilityFunctions';
 import { ImportGroup } from './ImportGroup';
 import { ImportGroupOrder } from './ImportGroupOrder';
 
 /**
  * Importgroup that processes all imports. Should be used if other groups don't process the import.
- * 
+ *
  * @export
  * @class RemainImportGroup
  * @implements {ImportGroup}
  */
 export class RemainImportGroup implements ImportGroup {
-    public readonly imports: Import[] = [];
+  public readonly imports: Import[] = [];
 
-    public get sortedImports(): Import[] {
-        const sorted = this.imports.sort((i1, i2) => importSort(i1, i2, this.order));
-        return [
-            ...sorted.filter(i => i instanceof StringImport),
-            ...sorted.filter(i => !(i instanceof StringImport)),
-        ];
-    }
+  public get sortedImports(): Import[] {
+    const sorted = this.imports.sort((i1, i2) => importSort(i1, i2, this.order));
+    return [
+      ...sorted.filter(i => i instanceof StringImport),
+      ...sorted.filter(i => !(i instanceof StringImport)),
+    ];
+  }
 
-    constructor(public readonly order: ImportGroupOrder = 'asc') { }
+  constructor(public readonly order: ImportGroupOrder = 'asc') { }
 
-    public reset(): void {
-        this.imports.length = 0;
-    }
+  public reset(): void {
+    this.imports.length = 0;
+  }
 
-    public processImport(tsImport: Import): boolean {
-        this.imports.push(tsImport);
-        return true;
-    }
+  public processImport(tsImport: Import): boolean {
+    this.imports.push(tsImport);
+    return true;
+  }
 }
