+++ src/import-organizer/index.ts
@@ -1,4 +1,4 @@
-import { inject, injectable } from 'inversify';
+import { inject, injectable, interfaces } from 'inversify';
 import { TypescriptParser } from 'typescript-parser';
 import { commands, ExtensionContext } from 'vscode';
 
@@ -6,6 +6,7 @@ import Activatable from '../activatable';
 import Configuration from '../configuration';
 import iocSymbols from '../ioc-symbols';
 import { Logger } from '../utilities/Logger';
+import ImportManager from './ImportManager';
 
 @injectable()
 export default class ImportOrganizer implements Activatable {
@@ -14,6 +15,7 @@ export default class ImportOrganizer implements Activatable {
     @inject(iocSymbols.logger) private logger: Logger,
     @inject(iocSymbols.configuration) private config: Configuration,
     @inject(iocSymbols.parser) private parser: TypescriptParser,
+    @inject(iocSymbols.importManager) private importManagerCtor: interfaces.Newable<ImportManager>,
   ) { }
 
   public setup(): void {
