+++ src/ioc-symbols.ts
@@ -2,6 +2,7 @@ export default {
   activatables: Symbol('activatables'),
   configuration: Symbol('configuration'),
   extensionContext: Symbol('extensionContext'),
+  importManager: Symbol('importManager'),
   logger: Symbol('logger'),
   parser: Symbol('parser'),
 };
